+++ sql/core/src/test/scala/org/apache/spark/sql/execution/vectorized/ColumnarBatchSuite.scala
@@ -31,13 +31,10 @@ import org.apache.spark.SparkFunSuite
 import org.apache.spark.memory.MemoryMode
 import org.apache.spark.sql.{RandomDataGenerator, Row}
 import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.catalyst.expressions.{SpecificInternalRow, UnsafeProjection}
 import org.apache.spark.sql.execution.arrow.ArrowUtils
-import org.apache.spark.sql.execution.columnar.ColumnAccessor
-import org.apache.spark.sql.execution.columnar.compression.ColumnBuilderHelper
 import org.apache.spark.sql.types._
 import org.apache.spark.unsafe.Platform
-import org.apache.spark.unsafe.types.{CalendarInterval, UTF8String}
+import org.apache.spark.unsafe.types.CalendarInterval
 
 class ColumnarBatchSuite extends SparkFunSuite {
 
@@ -1295,173 +1292,5 @@ class ColumnarBatchSuite extends SparkFunSuite {
 
     batch.close()
     allocator.close()
-  }
-
-  test("CachedBatch boolean Apis") {
-    val dataType = BooleanType
-    val columnBuilder = ColumnBuilderHelper(dataType, 1024, "col", true)
-    val row = new SpecificInternalRow(Array(dataType))
-
-    row.setNullAt(0)
-    columnBuilder.appendFrom(row, 0)
-    for (i <- 1 until 16) {
-      row.setBoolean(0, i % 2 == 0)
-      columnBuilder.appendFrom(row, 0)
-    }
-
-    val column = allocate(16, dataType, MemoryMode.ON_HEAP)
-    val columnAccessor = ColumnAccessor(dataType, columnBuilder.build)
-    ColumnAccessor.decompress(columnAccessor, column, 16)
-
-    assert(column.isNullAt(0) == true)
-    for (i <- 1 until 16) {
-      assert(column.isNullAt(i) == false)
-      assert(column.getBoolean(i) == (i % 2 == 0))
-    }
-    column.close
-  }
-
-  test("CachedBatch byte Apis") {
-    val dataType = ByteType
-    val columnBuilder = ColumnBuilderHelper(dataType, 1024, "col", true)
-    val row = new SpecificInternalRow(Array(dataType))
-
-    row.setNullAt(0)
-    columnBuilder.appendFrom(row, 0)
-    for (i <- 1 until 16) {
-      row.setByte(0, i.toByte)
-      columnBuilder.appendFrom(row, 0)
-    }
-
-    val column = allocate(16, dataType, MemoryMode.ON_HEAP)
-    val columnAccessor = ColumnAccessor(dataType, columnBuilder.build)
-    ColumnAccessor.decompress(columnAccessor, column, 16)
-
-    assert(column.isNullAt(0) == true)
-    for (i <- 1 until 16) {
-      assert(column.isNullAt(i) == false)
-      assert(column.getByte(i) == i)
-    }
-    column.close
-  }
-
-  test("CachedBatch short Apis") {
-    val dataType = ShortType
-    val columnBuilder = ColumnBuilderHelper(dataType, 1024, "col", true)
-    val row = new SpecificInternalRow(Array(dataType))
-
-    row.setNullAt(0)
-    columnBuilder.appendFrom(row, 0)
-    for (i <- 1 until 16) {
-      row.setShort(0, i.toShort)
-      columnBuilder.appendFrom(row, 0)
-    }
-
-    val column = allocate(16, dataType, MemoryMode.ON_HEAP)
-    val columnAccessor = ColumnAccessor(dataType, columnBuilder.build)
-    ColumnAccessor.decompress(columnAccessor, column, 16)
-
-    assert(column.isNullAt(0) == true)
-    for (i <- 1 until 16) {
-      assert(column.isNullAt(i) == false)
-      assert(column.getShort(i) == i)
-    }
-    column.close
-  }
-
-  test("CachedBatch int Apis") {
-    val dataType = IntegerType
-    val columnBuilder = ColumnBuilderHelper(dataType, 1024, "col", true)
-    val row = new SpecificInternalRow(Array(dataType))
-
-    row.setNullAt(0)
-    columnBuilder.appendFrom(row, 0)
-    for (i <- 1 until 16) {
-      row.setInt(0, i)
-      columnBuilder.appendFrom(row, 0)
-    }
-
-    val column = allocate(16, dataType, MemoryMode.ON_HEAP)
-    val columnAccessor = ColumnAccessor(dataType, columnBuilder.build)
-    ColumnAccessor.decompress(columnAccessor, column, 16)
-
-    assert(column.isNullAt(0) == true)
-    for (i <- 1 until 16) {
-      assert(column.isNullAt(i) == false)
-      assert(column.getInt(i) == i)
-    }
-    column.close
-  }
-
-  test("CachedBatch long Apis") {
-    val dataType = LongType
-    val columnBuilder = ColumnBuilderHelper(dataType, 1024, "col", true)
-    val row = new SpecificInternalRow(Array(dataType))
-
-    row.setNullAt(0)
-    columnBuilder.appendFrom(row, 0)
-    for (i <- 1 until 16) {
-      row.setLong(0, i.toLong)
-      columnBuilder.appendFrom(row, 0)
-    }
-
-    val column = allocate(16, dataType, MemoryMode.ON_HEAP)
-    val columnAccessor = ColumnAccessor(dataType, columnBuilder.build)
-    ColumnAccessor.decompress(columnAccessor, column, 16)
-
-    assert(column.isNullAt(0) == true)
-    for (i <- 1 until 16) {
-      assert(column.isNullAt(i) == false)
-      assert(column.getLong(i) == i.toLong)
-    }
-    column.close
-  }
-
-  test("CachedBatch float Apis") {
-    val dataType = FloatType
-    val columnBuilder = ColumnBuilderHelper(dataType, 1024, "col", true)
-    val row = new SpecificInternalRow(Array(dataType))
-
-    row.setNullAt(0)
-    columnBuilder.appendFrom(row, 0)
-    for (i <- 1 until 16) {
-      row.setFloat(0, i.toFloat)
-      columnBuilder.appendFrom(row, 0)
-    }
-
-    val column = allocate(16, dataType, MemoryMode.ON_HEAP)
-    val columnAccessor = ColumnAccessor(dataType, columnBuilder.build)
-    ColumnAccessor.decompress(columnAccessor, column, 16)
-
-    assert(column.isNullAt(0) == true)
-    for (i <- 1 until 16) {
-      assert(column.isNullAt(i) == false)
-      assert(column.getFloat(i) == i.toFloat)
-    }
-    column.close
-  }
-
-  test("CachedBatch double Apis") {
-    val dataType = DoubleType
-    val columnBuilder = ColumnBuilderHelper(dataType, 1024, "col", true)
-    val row = new SpecificInternalRow(Array(dataType))
-
-    row.setNullAt(0)
-    columnBuilder.appendFrom(row, 0)
-    for (i <- 1 until 16) {
-      row.setDouble(0, i.toDouble)
-      columnBuilder.appendFrom(row, 0)
-    }
-
-    val column = allocate(16, dataType, MemoryMode.ON_HEAP)
-    val columnAccessor = ColumnAccessor(dataType, columnBuilder.build)
-    ColumnAccessor.decompress(columnAccessor, column, 16)
-
-    assert(column.isNullAt(0) == true)
-    for (i <- 1 until 16) {
-      assert(column.isNullAt(i) == false)
-      assert(column.getDouble(i) == i.toDouble)
-    }
-    column.close
   }
 }
